♻️ Make Pf2eFrequency.getFrequency into a Pf2eJsonNodeReader default method
