+++ src/main/java/dev/ebullient/convert/tools/pf2e/Json2QuteFeat.java
@@ -32,7 +32,7 @@ public class Json2QuteFeat extends Json2QuteBase {
                 Field.alias.replaceTextFromList(rootNode, this),
                 Pf2eFeat.level.getTextOrDefault(rootNode, "1"),
                 Pf2eFeat.access.transformTextFrom(rootNode, ", ", this),
-                Pf2eFrequency.getFrequency(Pf2eFeat.frequency.getFrom(rootNode), this),
+                Pf2eFeat.frequency.getFrequencyFrom(rootNode, this),
                 Pf2eTypeReader.getQuteActivity(rootNode, Pf2eFeat.activity, this),
                 Pf2eFeat.trigger.transformTextFrom(rootNode, ", ", this),
                 Pf2eFeat.cost.transformTextFrom(rootNode, ", ", this),
@@ -62,7 +62,7 @@ public class Json2QuteFeat extends Json2QuteBase {
                 List.of(),
                 dedicationLevel,
                 Pf2eFeat.access.transformTextFrom(rootNode, ", ", this),
-                Pf2eFrequency.getFrequency(Pf2eFeat.frequency.getFrom(rootNode), this),
+                Pf2eFeat.frequency.getFrequencyFrom(rootNode, this),
                 Pf2eTypeReader.getQuteActivity(rootNode, Pf2eFeat.activity, this),
                 Pf2eFeat.trigger.transformTextFrom(rootNode, ", ", this),
                 Pf2eFeat.cost.transformTextFrom(rootNode, ", ", this),
